Replace NavBar setup
