+++ BeRich/AppDelegate.swift
@@ -1,0 +1,30 @@
+import SwiftUI
+
+final class AppDelegate: NSObject, UIApplicationDelegate {
+    func application(
+        _: UIApplication,
+        didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]? = nil
+    ) -> Bool {
+        setUpCustomNavBarAppearance()
+        return true
+    }
+}
+
+// MARK: - UI setUp
+
+extension AppDelegate {
+    private func setUpCustomNavBarAppearance() {
+        let backgroundColor = UIColor(Color.blueMain)
+        let titleColor = UIColor.white
+
+        let coloredAppearance = UINavigationBarAppearance()
+        coloredAppearance.configureWithTransparentBackground()
+        coloredAppearance.backgroundColor = backgroundColor
+        coloredAppearance.titleTextAttributes = [.foregroundColor: titleColor]
+        coloredAppearance.largeTitleTextAttributes = [.foregroundColor: titleColor]
+
+        UINavigationBar.appearance().standardAppearance = coloredAppearance
+        UINavigationBar.appearance().compactAppearance = coloredAppearance
+        UINavigationBar.appearance().scrollEdgeAppearance = coloredAppearance
+    }
+}
