+++ BeRich/BeRichApp.swift
@@ -2,6 +2,8 @@ import SwiftUI
 
 @main
 struct BeRichApp: App {
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+
     var body: some Scene {
         WindowGroup {
             ListScreen()
