--- BeRich/Modifiers/NavigationBarModifier.swift
@@ -1,42 +1,0 @@
-import SwiftUI
-
-private struct NavigationBarModifier: ViewModifier {
-    var backgroundColor: UIColor?
-    var titleColor: UIColor?
-
-    init(backgroundColor: UIColor?, titleColor: UIColor?) {
-        self.backgroundColor = backgroundColor
-        let coloredAppearance = UINavigationBarAppearance()
-        coloredAppearance.configureWithTransparentBackground()
-        coloredAppearance.backgroundColor = backgroundColor
-        coloredAppearance.titleTextAttributes = [.foregroundColor: titleColor ?? .white]
-        coloredAppearance.largeTitleTextAttributes = [.foregroundColor: titleColor ?? .white]
-
-        UINavigationBar.appearance().standardAppearance = coloredAppearance
-        UINavigationBar.appearance().compactAppearance = coloredAppearance
-        UINavigationBar.appearance().scrollEdgeAppearance = coloredAppearance
-    }
-
-    func body(content: Content) -> some View {
-        ZStack {
-            content
-            VStack {
-                GeometryReader { geometry in
-                    Color(backgroundColor ?? .clear)
-                        .frame(height: geometry.safeAreaInsets.top)
-                        .edgesIgnoringSafeArea(.top)
-                    Spacer()
-                }
-            }
-        }
-    }
-}
-
-extension View {
-    func navigationBarColor(backgroundColor: Color?, titleColor: Color?) -> some View {
-        modifier(NavigationBarModifier(
-            backgroundColor: backgroundColor.flatMap { UIColor($0) },
-            titleColor: titleColor.flatMap { UIColor($0) }
-        ))
-    }
-}
